[SNAP-4113](lib-gdal): Fix the GDAL distribution corruption issue when the GDALLoaderTest->testSyncGDALInitialisation() test, runs with the test data provided by "snap.reader.tests.data.dir".

+++ lib-gdal/src/test/java/org/esa/snap/dataio/gdal/GDALLoaderTest.java
@@ -111,8 +111,8 @@ public class GDALLoaderTest extends AbstractGDALTest {
         assertNotNull(TEST_GDAL_LOADER);
         try {
             Path gdalRootFolderPath = getExpectedNativeLibrariesRootFolderPath();
-            //remove gdal folder
-            if(Files.exists(gdalRootFolderPath)){
+            //remove gdal folder when not itialised to force the bundle deploy during this test
+            if(TEST_GDAL_LOADER.isNotInitialised() && Files.exists(gdalRootFolderPath)){
                 FileUtils.deleteTree(gdalRootFolderPath.toFile());
             }
